EC2 deploy

+++ JobTracker/src/main/java/com/skilldistillery/job/services/UserServiceImpl.java
@@ -10,14 +10,14 @@ import com.skilldistillery.job.entities.User;
 import com.skilldistillery.job.repositories.ApplicationRepository;
 import com.skilldistillery.job.repositories.UserRepository;
 
-
 @Service
 public class UserServiceImpl implements UserService {
-	
+
 	@Autowired
 	private UserRepository userRepo;
-	
-	@Autowired ApplicationRepository appliRepo;
+
+	@Autowired
+	ApplicationRepository appliRepo;
 
 	@Override
 	public List<User> findAllUsers() {
@@ -27,7 +27,7 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public User findByUserId(Integer id) {
 		Optional<User> user = userRepo.findById(id);
-		if(user.isPresent()) {
+		if (user.isPresent()) {
 			user.get();
 		}
 		return user.get();
@@ -35,7 +35,7 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public User createUser(User user) {
-		if(user != null) {
+		if (user != null) {
 			User createdUser = userRepo.saveAndFlush(user);
 		}
 		return user;
@@ -43,34 +43,34 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public User updateUserById(Integer id, User user) {
-		Optional<User> updatedUser = userRepo.findById(id); 
-		if(updatedUser.isPresent()) {
+		Optional<User> updatedUser = userRepo.findById(id);
+		if (updatedUser.isPresent()) {
 			User updateUser = updatedUser.get();
 			updateUser.setFirstName(user.getFirstName());
 
 			updateUser.setLastName(user.getLastName());
-			
+
 			updateUser.setEmail(user.getEmail());
-			
+
 			updateUser.setUsername(user.getUsername());
 			updateUser.setPassword(user.getPassword());
-			
-			
+
 			userRepo.saveAndFlush(updateUser);
-			
+
 		}
 		return user;
 	}
 
 	@Override
-	public User deleteUserById(Integer id) {
-		
+	public boolean deleteUserById(Integer id) {
+
 		Optional<User> existingUser = userRepo.findById(id);
-		if(existingUser.isPresent()) {
+		if (existingUser.isPresent() && existingUser.get().isEnabled() == true) {
 			userRepo.deleteById(id);
-			
+			return true;
+		} else {
+			return false;
 		}
-		return null;
 	}
 
 }
